Version 5.1.0,  Revision 21733,  Hash e1c961fa5
MODIFIED *  config/mk/global/targets.mk driver/ypp.F include/driver/version.h sbin/compilation/object_remove.sh sbin/compilation/object_save_and_restore.sh ypp/modules/NL_project.dep ypp/modules/YPP_ELPH_project.dep ypp/modules/YPP_RT_project.dep ypp/modules/mod_YPP_ELPH.F ypp/modules/mod_YPP_real_time.F

NEW *  driver/YPP_SC_project.dep

Bugs:
- [configure] Few more fixes to projects

Patch sent by:  Davide Sangalli <davide.sangalli@gmail.com>

+++ sbin/compilation/object_save_and_restore.sh
@@ -60,7 +60,8 @@ if [[ -f $dir/$restore_dir/$library ]] ; then
  count_mods=`ls -1 $dir/*.mod 2>/dev/null | wc -l`
  count_modr=`ls -1 $dir/$restore_dir/*.mod 2>/dev/null | wc -l`
  count_f90=`ls -1 $dir/$restore_dir/*.f90 2>/dev/null | wc -l`
- rm $dir/*.o
+ count_obj=`ls -1 $dir/*.o 2>/dev/null | wc -l`
+ if [ $count != 0 ]; then rm $dir/*.o  ; fi
  cp $dir/$restore_dir/*.o $dir/ ;
  if [ "$count_mods" -gt "0" ] ; then
   cd $dir
